+++ crates/nu-command/src/core_commands/mod.rs
@@ -20,6 +20,7 @@ mod export_use;
 mod extern_;
 mod for_;
 pub mod help;
+pub mod help_commands;
 mod help_operators;
 mod hide;
 mod hide_env;
@@ -59,6 +60,7 @@ pub use export_use::ExportUse;
 pub use extern_::Extern;
 pub use for_::For;
 pub use help::Help;
+pub use help_commands::HelpCommands;
 pub use help_operators::HelpOperators;
 pub use hide::Hide;
 pub use hide_env::HideEnv;
