+++ crates/nu-command/src/default_context.rs
@@ -50,6 +50,7 @@ pub fn create_default_context() -> EngineState {
             Extern,
             For,
             Help,
+            HelpCommands,
             HelpOperators,
             Hide,
             HideEnv,
